Ok, it was late last night, but now brain is functioning and found that I was importing using require instead of artifacts.require.

+++ contracts/JAF.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.2;
 contract JAF {
     string public name = "JAF";
     string public symbol = "JAF";
-    uint256 public totalSupply = 1000000000000000000000000; // 1 million tokens
+    uint public totalSupply = 1000000000000000000000000; // 1 million tokens
     uint8 public decimals = 18;
 
     event Transfer(address indexed _from, address indexed _to, uint _value);
@@ -15,17 +15,14 @@ contract JAF {
         uint _value
     );
 
-    mapping(address => uint256) public balanceOf;
-    mapping(address => mapping(address => uint256)) public allowance;
+    mapping(address => uint) public balanceOf;
+    mapping(address => mapping(address => uint)) public allowance;
 
     constructor() {
         balanceOf[msg.sender] = totalSupply;
     }
 
-    function transfer(
-        address _to,
-        uint256 _value
-    ) public returns (bool success) {
+    function transfer(address _to, uint _value) public returns (bool success) {
         require(balanceOf[msg.sender] >= _value);
         balanceOf[msg.sender] -= _value;
         balanceOf[_to] += _value;
@@ -35,7 +32,7 @@ contract JAF {
 
     function approve(
         address _spender,
-        uint256 _value
+        uint _value
     ) public returns (bool success) {
         allowance[msg.sender][_spender] = _value;
         emit Approval(msg.sender, _spender, _value);
@@ -45,7 +42,7 @@ contract JAF {
     function transferFrom(
         address _from,
         address _to,
-        uint256 _value
+        uint _value
     ) public returns (bool success) {
         require(_value <= balanceOf[_from]);
         require(_value <= allowance[_from][msg.sender]);
