+++ contracts/Tether.sol
@@ -4,27 +4,24 @@ pragma solidity ^0.8.2;
 contract Tether {
     string public name = "Mock Tether";
     string public symbol = "mUSDT";
-    uint256 public totalSupply = 1000000000000000000000000; // 1 million tokens
+    uint public totalSupply = 1000000000000000000000000; // 1 million tokens
     uint8 public decimals = 18;
 
-    event Transfer(address indexed _from, address indexed _to, uint256 _value);
+    event Transfer(address indexed _from, address indexed _to, uint _value);
     event Approval(
         address indexed _owner,
         address indexed _spender,
-        uint256 _value
+        uint _value
     );
 
-    mapping(address => uint256) public balanceOf;
-    mapping(address => mapping(address => uint256)) public allowance;
+    mapping(address => uint) public balanceOf;
+    mapping(address => mapping(address => uint)) public allowance;
 
     constructor() {
         balanceOf[msg.sender] = totalSupply;
     }
 
-    function transfer(
-        address _to,
-        uint256 _value
-    ) public returns (bool success) {
+    function transfer(address _to, uint _value) public returns (bool success) {
         require(balanceOf[msg.sender] >= _value);
         balanceOf[msg.sender] -= _value;
         balanceOf[_to] += _value;
@@ -34,7 +31,7 @@ contract Tether {
 
     function approve(
         address _spender,
-        uint256 _value
+        uint _value
     ) public returns (bool success) {
         allowance[msg.sender][_spender] = _value;
         emit Approval(msg.sender, _spender, _value);
@@ -44,7 +41,7 @@ contract Tether {
     function transferFrom(
         address _from,
         address _to,
-        uint256 _value
+        uint _value
     ) public returns (bool success) {
         require(_value <= balanceOf[_from]);
         require(_value <= allowance[_from][msg.sender]);
